Relocate and update deployment configuration file
The file previously known as 'asa-performance-tests.yml' has been renamed to 'deployment.yml' and relocated to the 'resources' directory. Along with the renaming, the app name and image have been updated to 'java-app' and 'joaonunes96/argocd-test:1.0' respectively, with a defined container port of 8080. Unnecessary copy and command directives in the Dockerfile have also been removed.