Change Deployment to Job in deployment.yml
In the resources/deployment.yml, the Kubernetes Deployment has been switched to a Job. In addition to the change in the kind of the resource, the container now executes a new command, and Job-specific fields like 'parallelism' and 'completions' have been added. The restart policy has also been set to 'Never'.